ask-server/rag/ocr/requirements.txt
pymupdf
pytesseract
pillow

# sudo apt install tesseract-ocr